feat/SV-180: 사용하지 않는 import 삭제
